feat: core types + arithmetic security hardening (PR 1/7)
New types and fields:
- Add MintData, GenesisSpend types to clvm_zk_core/types.rs
- Add ProofType::Mint = 3 variant to structures.rs; included in is_submittable()
- CoinMode enum (Execute | Spend | Mint) replaces serial_commitment_data/mint_data Options

Security: compile-time mutual exclusion (spend vs mint):
- CoinMode::Spend + CoinMode::Mint are now separate variants, not two Option fields
- Spend+Mint combination is a compile error, not a runtime panic

Security: enforce_ring_balance explicit Mint rejection (F-01):
- CoinMode::Mint arm returns Err(...) rather than falling through with input=0
- Prevents future Mint inputs from bypassing balance enforcement

Security: host-side guards before zkVM guests (F-02):
- risc0/src/lib.rs + sp1/src/lib.rs: reject CoinMode::Mint with clean Err before
  submitting to prover — avoids opaque guest panic and DoS of proof queue

Security: tail_source doc warning (F-03):
- Removed #[serde(default)] from tail_source; doc explicitly warns that enforcement
  must require Some(_) for delta-negative CAT spends, not treat None as safe default

Security: leaf_index arch-safe serialization:
- leaf_index: usize → u64 in SerialCommitmentData, GenesisSpend (Borsh wire format)
- TakerCoinData.leaf_index: usize → u64 in both settlement guests + settlement.rs
  local structs (F-07: previously missed in the migration)
- All verify_merkle_proof callsites: leaf_index as usize → try_from().expect() (F-05)
  replaces silent truncation with legible abort message

Security: arithmetic hardening:
- enforce_ring_balance: + → checked_add; CoinMode::Execute arm explicit
- modular_pow: zero-modulus guard; doc comment; two unit tests (F-06)

All 144 mock + 31 clvm_zk_core tests pass.

+++ src/protocol/settlement.rs
@@ -204,7 +204,8 @@ pub fn prove_settlement(params: SettlementParams) -> Result<SettlementProof, Pro
             serial_number: [u8; 32],
             serial_randomness: [u8; 32],
             merkle_path: Vec<[u8; 32]>,
-            leaf_index: usize,
+            /// u64 (not usize) — must match guest TakerCoinData for consistent serde on 32-bit zkVM
+            leaf_index: u64,
         }
 
         let input = SettlementInput {
@@ -220,7 +221,7 @@ pub fn prove_settlement(params: SettlementParams) -> Result<SettlementProof, Pro
                 serial_number: params.taker_secrets.serial_number,
                 serial_randomness: params.taker_secrets.serial_randomness,
                 merkle_path: params.taker_merkle_path,
-                leaf_index: params.taker_leaf_index,
+                leaf_index: params.taker_leaf_index as u64,
             },
             merkle_root: params.merkle_root,
             payment_nonce: params.payment_nonce,
@@ -334,7 +335,8 @@ pub fn prove_settlement(params: SettlementParams) -> Result<SettlementProof, Pro
             serial_number: [u8; 32],
             serial_randomness: [u8; 32],
             merkle_path: Vec<[u8; 32]>,
-            leaf_index: usize,
+            /// u64 (not usize) — must match guest TakerCoinData for consistent serde on 32-bit zkVM
+            leaf_index: u64,
         }
 
         let input = SettlementInput {
@@ -350,7 +352,7 @@ pub fn prove_settlement(params: SettlementParams) -> Result<SettlementProof, Pro
                 serial_number: params.taker_secrets.serial_number,
                 serial_randomness: params.taker_secrets.serial_randomness,
                 merkle_path: params.taker_merkle_path,
-                leaf_index: params.taker_leaf_index,
+                leaf_index: params.taker_leaf_index as u64,
             },
             merkle_root: params.merkle_root,
             payment_nonce: params.payment_nonce,
